+++ ChangeLog.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Fixed
+
+- fixed a bug in derived queries when using two times `@Relations` in one entity
+
 ## [1.3.0] - 2018-07-18
 
 ### Deprecated